Work on logger
